requirements.txt
accelerate==0.17.1
addict==2.4.0
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.2
ConfigArgParse==1.5.3
contourpy==1.0.7
cycler==0.11.0
dash==2.9.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.6
decorator==5.1.1
diffusers==0.14.0
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.10.0
Flask==2.2.3
fonttools==4.39.2
huggingface-hub==0.13.3
idna==3.4
imageio==2.26.1
importlib-metadata==6.1.0
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==8.11.0
ipywidgets==8.0.4
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.0
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.9.4.25
open3d==0.17.0
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.1
plotly==5.13.1
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
pymeshlab==2022.2.post3
pyparsing==3.0.9
pyquaternion==0.9.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.2
regex==2022.10.31
requests==2.28.2
Rtree==1.0.1
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
stack-data==0.6.2
tenacity==8.2.2
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.27.2
trimesh==3.20.2
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.2.3
widgetsnbextension==4.0.5
zipp==3.15.0
